failed attempt at zooming directly
This could work, but the problem is that even though the card does
properly zoom and unzoom correctly, it is impossible to put it into the
foreground. toFront() and toBack() only work on the Group level. Since
the CardView and its count label are both in a Group together, this
won’t work.

An alternative approach will be to simply have another layer, and when
a long press is detected, populate the card data and render that object
in the center of the screen.