basic card zoom working in Deck Builder
Works with double click or with long press, with a slight problem.
The problem is that when you double click, the first click still
registers as a single click, possibly adding the card to your current
deck list while you are trying to double click to zoom.
There may be a workaround, keeping track of when the first click is
registered, and not allowing it to be a single click until after the
time limit for double click, but i don’t know of a good way to add a
timer to a ClickListener that will be compatible with GWT.
Another possible workaround would be to roll back a single click if a
double click is detected, but currently all that is happening is a call
to addEntity(), and doing something like removeEntity() might have
unexpected consequences.