changed zoom a bit, now working for GameScreen
there is a problem here. Since the callback of the CardViewSmall is the
DefaultZoneView, and the GameScreen contains more than one of these,
each one of them is allowed to open up one zoomed in card, and thus
multiple cards can be zoomed in at the same time.
The solution could be for DefaultZoneView to mark a boolean in the
GameScreen as true when a card is zoomed in, and check against that
when zooming another card. Alternatively, the zoom operation could be
passed along to the GameScreen.